Updated profiles with approximate EA3 options
Case ZC-887:  We needed to add our newly-created mcrypt extension to all
profiles.  While we were at it, we want to provide as similar a set of
profiles between EA3 and EA4 that we can, so also added the rest of the
packages which EA3 provides, where possible.

Also, whitespace cleanup and ABCed one of the profiles.